Voice-Assistant-master copy/backend/requirements.txt
flask
flask-cors
python-multipart
groq
sarvamai
python-dotenv